+++ .env
@@ -1,3 +1,3 @@
 EXPO_PUBLIC_SUPABASE_URL=
 EXPO_PUBLIC_SUPABASE_ANON_KEY=
-EXPO_PUBLIC_CLERK_PUBLISHABLE_KEY=pk_test_dGFsZW50ZWQtZmlyZWZseS04Mi5jbGVyay5hY2NvdW50cy5kZXYk
+EXPO_PUBLIC_CLERK_PUBLISHABLE_KEY=
