+++ app/(auth)/Journal.tsx
