+++ app/(auth)/Record.tsx
